G2T2-19 Chore: Re-organise files
