+++ client/src/App.tsx
@@ -10,6 +10,7 @@ import "./styles/App.css";
 import logo from "./assets/logo.png";
 import { BrowserRouter, Link, Route, Routes } from "react-router-dom";
 import Home from "./pages/Home";
+import LearningJourney from "./pages/LearningJourney";
 import Roles from "./pages/Roles";
 import Skills from "./pages/Skills";
 import Courses from "./pages/Courses";
@@ -72,8 +73,9 @@ const App = () => (
           Eric
         </Button>
         <Routes>
-          <Route path="/" element={<Home />} />
-          <Route path="/roles" element={<Roles lj={true} />} />
+          <Route path="/" element={<Home lj={true}/>} />
+          <Route path="/LearningJourney" element={<LearningJourney />} />
+          <Route path="/roles" element={<Roles />} />
           <Route path="/skills" element={<Skills />} />
           <Route path="/courses" element={<Courses />} />
           <Route path="/create/2" element={<Home />} />
